UnitOFWork 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/Message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,7 +16,7 @@ export class MessagesComponent implements OnInit {
   pageNumber=1;
   pageSize=5;
   laoding=false;
-  constructor(private messageSerrvice:MessageService) { }
+  constructor(private messageSerrvice:MessageService,private confirmService:ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -37,10 +38,15 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id:number){
-    this.messageSerrvice.deleteMessage(id).subscribe(()=>{
-     
-       this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
-    });
+    this.confirmService.confirm('Confirm delete message','this cannot be undone')
+        .subscribe(result=>{
+          if(result){
+            this.messageSerrvice.deleteMessage(id).subscribe(()=>{
+              this.messages.splice(this.messages.findIndex(m=>m.id===id),1);
+            });
+          }
+        });
+       
   }
 
 }
